requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn[standard]>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
sqlalchemy>=1.4.23,<1.5.0
psycopg2-binary>=2.9.1,<3.0.0
pydantic>=1.8.2,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=0.19.0,<0.20.0
jinja2>=3.0.1,<4.0.0
aiofiles>=0.7.0,<0.8.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0
sentencepiece>=0.1.99,<0.2.0
accelerate>=0.20.0,<0.21.0
email-validator>=1.1.3,<2.0.0 